Lint

+++ app/Models/Profession.ts
@@ -34,10 +34,8 @@ export default class Profession extends compose(BaseModel, SoftDeletes) {
   public static async deleteServices(profession: Profession) {
     const services = await Service.query().where('profession_id', profession.id)
     for (let service of services) {
-      if (service.picto_file)
-        await UploadService.deleteFile(service.picto_file)
-      if (service.image_file)
-        await UploadService.deleteFile(service.image_file)
+      if (service.picto_file) await UploadService.deleteFile(service.picto_file)
+      if (service.image_file) await UploadService.deleteFile(service.image_file)
       await service.delete()
     }
   }
